Moved Category and CategoryDao to the 'category' package
