+++ app/src/main/java/fr/free/nrw/commons/category/CategorizationFragment.java
@@ -35,8 +35,6 @@ import butterknife.BindView;
 import butterknife.ButterKnife;
 import dagger.android.support.DaggerFragment;
 import fr.free.nrw.commons.R;
-import fr.free.nrw.commons.data.Category;
-import fr.free.nrw.commons.data.CategoryDao;
 import fr.free.nrw.commons.mwapi.MediaWikiApi;
 import fr.free.nrw.commons.upload.MwVolleyApi;
 import fr.free.nrw.commons.utils.StringSortingUtils;
